move admin layout under "views/admin/" folder
