+++ src/views/admin/Questionnaire/QuestionnaireList.vue
@@ -1,7 +1,7 @@
 <script setup>
 import { ref, computed } from 'vue'
 import { useRouter } from 'vue-router'
-import AdminLayout from '@/layouts/AdminLayout.vue'
+import AdminLayout from '@/views/admin/AdminLayout.vue'
 import { eventsMockData } from '@/mock/eventsMock.js'
 
 // Get the router object
