+++ src/views/admin/User/UserEdit.vue
@@ -1,7 +1,7 @@
 <script setup>
 import { ref, onMounted } from "vue";
 import { useRoute, useRouter } from "vue-router";
-import AdminLayout from "@/layouts/AdminLayout.vue";
+import AdminLayout from "@/views/admin/AdminLayout.vue";
 import { usersMockData } from "@/mock/usersMock.js";
 import { getUserById, updateUser } from "@/api/userServices";
