+++ src/views/admin/User/UserList.vue
@@ -1,7 +1,7 @@
 <script setup>
 import { ref, computed, onMounted } from 'vue'
 import { useRouter } from 'vue-router'
-import AdminLayout from '@/layouts/AdminLayout.vue'
+import AdminLayout from '@/views/admin/AdminLayout.vue'
 import { usersMockData } from '@/mock/usersMock.js'
 import { fetchAllUsersData,deleteUserbyID } from '@/api/userServices'
